Implement map drawing using Ancillary Text Record
On GOES-R satellites, the data contained in the Image
Navigation Header is not completely accurate. There
is an issue with round-off which causes a minor offset.
For this reason alone, better navigation data is desired.

Additionally, there appears to be a further error in
NOAA's computation of the Image Navigation Header such
that COFF, LOFF, CFAC, and LFAC are often slightly off
from what they should be. The difference can vary
between channels of different scales (ex. Mesoscale).
NOAA is aware of the issue, but they have not found a
solution as of this writing.

To solve both of these problems, let's use some of the
additional navigation data available in the Ancillary
Text Header, if it is available. In these cases, the
old "magic constant" is no longer required, and the
resulting projection is more accurate. Currently, the
ancillary text record can be used for all GOES FD
and Mesoscale imagery, including non-CMIP imagery.

Any data missing navigation data in the ancillary text
record, including old data and the Himawari relay on
GOES-18, fall back to using the Image Navigation Record.
In this case, the "magic constant" has been updated to
a more accurate value. See the new code comment in
map_drawer.cc for its derivation.